+++ src/Symfony/Component/Form/Extension/Core/DataTransformer/DateTimeToHtml5LocalDateTimeTransformer.php
@@ -23,6 +23,12 @@ use Symfony\Component\Form\Exception\TransformationFailedException;
 class DateTimeToHtml5LocalDateTimeTransformer extends BaseDateTimeTransformer
 {
     public const HTML5_FORMAT = 'Y-m-d\\TH:i:s';
+    public const HTML5_WITHOUT_SECONDS_FORMAT = 'Y-m-d\\TH:i';
+
+    public function __construct(string $inputTimezone = null, string $outputTimezone = null, private readonly bool $withSeconds = true)
+    {
+        parent::__construct($inputTimezone, $outputTimezone);
+    }
 
     /**
      * Transforms a \DateTime into a local date and time string.
@@ -54,7 +60,7 @@ class DateTimeToHtml5LocalDateTimeTransformer extends BaseDateTimeTransformer
             $dateTime = $dateTime->setTimezone(new \DateTimeZone($this->outputTimezone));
         }
 
-        return $dateTime->format(self::HTML5_FORMAT);
+        return $dateTime->format($this->withSeconds ? self::HTML5_FORMAT : self::HTML5_WITHOUT_SECONDS_FORMAT);
     }
 
     /**